requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pyyaml>=5.4.0
ultralytics>=8.0.0
fastapi
uvicorn
requests
python-multipart
pydantic>=1.10.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
aiofiles>=0.8.0
aiohttp>=3.8.0
pydantic-settings>=2.0.0
lap>=0.5.12
